requirements.txt
install
requests
beautifulsoup4
langchain
openai
faiss-cpu
toktoken
streamlit
python-dotenv
pypdf
pandas
streamlit
